+++ test/deno_entry.js
@@ -14,6 +14,8 @@ import "./domain/entities/GitTreeBuilder.test.js";
 import "./domain/entities/GitTreeEntry.test.js";
 import "./domain/errors/Errors.test.js";
 import "./domain/services/ByteMeasurer.test.js";
+import "./domain/services/CommandSanitizer.test.js";
 import "./domain/services/EnvironmentPolicy.test.js";
+import "./domain/services/GitErrorClassifier.test.js";
 import "./domain/value-objects/GitFileMode.test.js";
 import "./domain/value-objects/GitObjectType.test.js";